+++ include/ALGraphV2.h
@@ -12,13 +12,13 @@
 // Adjacency list implementation of Graph
 class ALGraphV2{
     uint64_t v, e;
-    std::vector<uint64_t>* edges;
-    std::vector<uint64_t>* weights;
+    std::vector<int>* edges;
+    std::vector<int>* weights;
 
 public:
     ALGraphV2(uint64_t v, uint64_t e) : v(v), e(e){
-        edges = new std::vector<uint64_t>[v + 2];
-        weights = new std::vector<uint64_t>[v + 2];
+        edges = new std::vector<int>[v + 2];
+        weights = new std::vector<int>[v + 2];
         for(int i = 0; i <= v; ++i)
             edges[i].clear(), weights[i].clear();
     }
@@ -39,15 +39,15 @@ public:
 
     void sortByEdgesByNodeId();
 
-    inline std::vector<uint64_t>::iterator begin(int cur_vertex) {
+    inline std::vector<int>::iterator begin(int cur_vertex) {
         return edges[cur_vertex].begin();
     }
 
-    inline std::vector<uint64_t>::iterator end(int cur_vertex){
+    inline std::vector<int>::iterator end(int cur_vertex){
         return edges[cur_vertex].end();
     }
 
-    inline std::vector<uint64_t>::iterator begin_weights(int cur_vertex) {
+    inline std::vector<int>::iterator begin_weights(int cur_vertex) {
         return weights[cur_vertex].begin();
     }
 };
