+++ src/CSRGraph.cpp
@@ -25,18 +25,18 @@ void CSRGraph::finished() {
 
 void CSRGraph::add_edge(int from, int to, int weight) {
     // adding a new edge
-    delta_e = new uint64_t[e+1];
-    memcpy(delta_e, edges, sizeof(uint64_t) * offsets[from + 1]);
+    delta_e = new int[e+1];
+    memcpy(delta_e, edges, sizeof(int) * offsets[from + 1]);
     delta_e[offsets[from + 1]] = to;
-    memcpy(delta_e + offsets[from + 1] + 1, edges + offsets[from + 1], sizeof(uint64_t) * (e - offsets[from + 1]));
+    memcpy(delta_e + offsets[from + 1] + 1, edges + offsets[from + 1], sizeof(int) * (e - offsets[from + 1]));
     delete[] edges;
     edges = delta_e;
 
     // adding weight for the new edge
-    delta_w = new uint64_t[e+1];
-    memcpy(delta_w, weights, sizeof(uint64_t) * offsets[from + 1]);
+    delta_w = new int[e+1];
+    memcpy(delta_w, weights, sizeof(int) * offsets[from + 1]);
     delta_w[offsets[from + 1]] = weight;
-    memcpy(delta_w + offsets[from + 1] + 1, weights + offsets[from + 1], sizeof(uint64_t) * (e - offsets[from + 1]));
+    memcpy(delta_w + offsets[from + 1] + 1, weights + offsets[from + 1], sizeof(int) * (e - offsets[from + 1]));
     delete[] weights;
     weights = delta_w;
